PedSearch/requirements.txt
faiss-cpu
sentence-transformers
numpy
flask
openai
gunicorn
python-dotenv